Add site level integration extensions
This is done by the creation of Binder classes that are then added to a
list in settings.py called WIKI_BINDERS. We could probably do something
fancy like auto load all classes that inherit from wiki.binders.Binder
through a Metaclass registry or something but I think I prefer it to be
explicit via a setting for now.

The Binder is responsible for URL patterns, URL reversals, and object
lookups.